move tracking to config
