+++ cmd/flow/main.go
@@ -33,7 +33,6 @@ import (
 	"github.com/onflow/flow-cli/internal/project"
 	"github.com/onflow/flow-cli/internal/quick"
 	"github.com/onflow/flow-cli/internal/scripts"
-	"github.com/onflow/flow-cli/internal/settings"
 	"github.com/onflow/flow-cli/internal/signatures"
 	"github.com/onflow/flow-cli/internal/snapshot"
 	"github.com/onflow/flow-cli/internal/status"
@@ -59,7 +58,6 @@ func main() {
 	// single commands
 	status.Command.AddToParent(cmd)
 	tools.DevWallet.AddToParent(cmd)
-	settings.TrackingSettings.AddToParent(cmd)
 
 	// structured commands
 	cmd.AddCommand(cadence.Cmd)
