+++ internal/config/config.go
@@ -32,6 +32,7 @@ func init() {
 	InitCommand.AddToParent(Cmd)
 	Cmd.AddCommand(AddCmd)
 	Cmd.AddCommand(RemoveCmd)
+	TrackingSettings.AddToParent(Cmd)
 }
 
 type Result struct {
